feat(mime): ready for review
Refs: #1258
